refactor(ui): break import cycle
